requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.31
psycopg2-binary>=2.9.9
pandas>=2.2.2
pydantic>=2.7.4
python-dotenv>=1.0.1
gspread>=6.0.0
google-auth>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
